Parser signature: remove offset parameter
The offset parameter exists to prevent allocating substrings from all of
the parsers as we go. Instead we can reference a slot in the original
string and only pass around those small pointers.

This is an optimization that isn't warranted yet and is confusing the
comprehension of the code. It may turn out to be a bad optimization when
considering global optimization when parsing thousands of addresses.

Also included are stripping whitespace on the boundaries of function
calls and array indices. This is to become more idiomatic.